refactor: move KSName into common-util
