+++ common-util/src/main/kotlin/com/google/devtools/ksp/KSPUtils.kt
@@ -1,5 +1,8 @@
 package com.google.devtools.ksp
 
+import com.google.devtools.ksp.processing.impl.KSNameImpl
+import org.jetbrains.kotlin.name.ClassId
+
 class IdKey<T>(private val k: T) {
     override fun equals(other: Any?): Boolean = if (other is IdKey<*>) k === other.k else false
     override fun hashCode(): Int = k.hashCode()
@@ -16,3 +19,5 @@ class IdKeyTriple<T, P, Q>(private val k1: T, private val k2: P, private val k3:
         k2 === other.k2 && k3 === other.k3 else false
     override fun hashCode(): Int = k1.hashCode() * 31 * 31 + k2.hashCode() * 31 + k3.hashCode()
 }
+
+fun ClassId.toKSName() = KSNameImpl.getCached(asSingleFqName().toString())
