+++ common-util/src/main/kotlin/com/google/devtools/ksp/processing/impl/KSNameImpl.kt
@@ -1,0 +1,22 @@
+package com.google.devtools.ksp.processing.impl
+
+import com.google.devtools.ksp.KSObjectCache
+import com.google.devtools.ksp.symbol.KSName
+
+class KSNameImpl private constructor(val name: String) : KSName {
+    companion object : KSObjectCache<String, KSNameImpl>() {
+        fun getCached(name: String) = cache.getOrPut(name) { KSNameImpl(name) }
+    }
+
+    override fun asString(): String {
+        return name
+    }
+
+    override fun getQualifier(): String {
+        return name.split(".").dropLast(1).joinToString(".")
+    }
+
+    override fun getShortName(): String {
+        return name.split(".").last()
+    }
+}
