+++ compiler-plugin/src/main/kotlin/com/google/devtools/ksp/symbol/impl/binary/KSTypeAliasDescriptorImpl.kt
@@ -1,8 +1,8 @@
 package com.google.devtools.ksp.symbol.impl.binary
 
 import com.google.devtools.ksp.KSObjectCache
+import com.google.devtools.ksp.processing.impl.KSNameImpl
 import com.google.devtools.ksp.symbol.*
-import com.google.devtools.ksp.symbol.impl.kotlin.KSNameImpl
 import com.google.devtools.ksp.toKSModifiers
 import org.jetbrains.kotlin.descriptors.TypeAliasDescriptor
