--- compiler-plugin/src/main/kotlin/com/google/devtools/ksp/symbol/impl/kotlin/KSNameImpl.kt
@@ -1,39 +1,0 @@
-/*
- * Copyright 2020 Google LLC
- * Copyright 2010-2020 JetBrains s.r.o. and Kotlin Programming Language contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.devtools.ksp.symbol.impl.kotlin
-
-import com.google.devtools.ksp.KSObjectCache
-import com.google.devtools.ksp.symbol.KSName
-
-class KSNameImpl private constructor(val name: String) : KSName {
-    companion object : KSObjectCache<String, KSNameImpl>() {
-        fun getCached(name: String) = cache.getOrPut(name) { KSNameImpl(name) }
-    }
-
-    override fun asString(): String {
-        return name
-    }
-
-    override fun getQualifier(): String {
-        return name.split(".").dropLast(1).joinToString(".")
-    }
-
-    override fun getShortName(): String {
-        return name.split(".").last()
-    }
-}
