+++ compiler-plugin/src/main/kotlin/com/google/devtools/ksp/symbol/impl/synthetic/KSConstructorSyntheticImpl.kt
@@ -19,9 +19,9 @@ package com.google.devtools.ksp.symbol.impl.synthetic
 
 import com.google.devtools.ksp.KSObjectCache
 import com.google.devtools.ksp.isPublic
+import com.google.devtools.ksp.processing.impl.KSNameImpl
 import com.google.devtools.ksp.processing.impl.ResolverImpl
 import com.google.devtools.ksp.symbol.*
-import com.google.devtools.ksp.symbol.impl.kotlin.KSNameImpl
 
 class KSConstructorSyntheticImpl private constructor(val ksClassDeclaration: KSClassDeclaration) :
     KSFunctionDeclaration,
