+++ compiler-plugin/src/main/kotlin/com/google/devtools/ksp/symbol/impl/synthetic/KSErrorTypeClassDeclaration.kt
@@ -17,9 +17,9 @@
 
 package com.google.devtools.ksp.symbol.impl.synthetic
 
+import com.google.devtools.ksp.processing.impl.KSNameImpl
 import com.google.devtools.ksp.processing.impl.ResolverImpl
 import com.google.devtools.ksp.symbol.*
-import com.google.devtools.ksp.symbol.impl.kotlin.KSNameImpl
 
 object KSErrorTypeClassDeclaration : KSClassDeclaration {
     override val annotations: Sequence<KSAnnotation> = emptySequence()
